Functional 1.5 build.
Necessary changes have been made for the assembly to be buildable against the 1.5 version. More work will be needed to verify everything actually functions.
- Refactor CostToMoveIntoCell from int to float.
- Add RimWorld dependency to IngredientValueGetter_Mass
- Disable breaking sound for RangedShieldBelt

+++ Source/communityframework/communityframework/Harmony patches/CostToMoveIntoCell.cs
@@ -19,7 +19,7 @@ namespace CF
         [HarmonyPatch(new Type[] { typeof(Pawn), typeof(IntVec3) })]
         class CostToMoveIntoCell
         {
-            static int Postfix(int __result, Pawn pawn, IntVec3 c)
+            static float Postfix(float __result, Pawn pawn, IntVec3 c)
             {
                 if (pawn.health.hediffSet.GetAllComps().Where(hediff => { return hediff is HediffComp_IgnorePathCost; }).Any())
                 {
